Add read_all() function which swallows eof errors
read_all() does what it says on the tin: it reads the entirity of
a stream into a dynamic buffer. End-of-file is not considered an
error in this case, and so is swallowed up.

At the moment you can only pass dynamic buffers, but it should
work for static buffers too. The only consideration there is what
to do if the buffer is too small to read everything: we'll have
to decide what sort of error to return.