[travis] Install a recent CMake on Linux
The Kalakris repo only contains CMake 2.8, and we require 3.1
Also, don't bother using Ninja as apparently the downloaded
CMake doesn't like it

Apparently the downloaded CMake doesn't like it, so let's not
bother. It's not like the CI checks are that fast anyway, so Ninja
won't make a huge difference.