Add mmap tests to read benchmark
The verdict: an mmap’d file is a tad slower than a direct file read,
but there’s not much in it.

A range read on an mmap’d file (using the RandomAccess specialisation)
is about half the speed of a direct read, which is a bit sad. It’s
still about 1/3 faster than a non-RA read though.