Add a locking stream adaptor
This allows you to wrap any stream (well, currently *anything*: the
type is unconstrained) in a type that contains the stream along with a
mutex. Calling the `lock()` member function returns a stream adaptor on
which you can do actual stream operations.

This seems like quite an elegant solution to the problem of
multithreaded file locking, but this stuff is always tricky. Hopefully
there aren’t too many bugs.