Add std::[experimental]::filesystem include wrapper
If <filesystem> is available, this uses it; otherwise it tries
<experimental/filesystem>. If either of those are available it brings
filesystem::path and filesystem::perms (the only bits we currently use)
into the io_std::filesystem namespace.

If neither header is found (with libc++ for example) then we include an
implementation of fs::perms, and fs::path becomes an alias for
std::string.