Rewrite string_stream
string_stream is now templated, so it can read from any of the four
varieties of std::string. It also uses the memory_stream read and seek
implementation.

Also, I really need to think of a good name of “a Seeable
SyncReadStream which is just a chunk of memory”, because that’s really
a concept of its own and is quite important, because it’s something we
can use random access iterators on (rather than just input iterators
like normal streams).