Add open_mode.hpp
This header contains an enum class which defines the various open flags
that can be passed to a file::open() method. These are broadly
equivalent to the Posix O_* flags that are passed to open(2) (and
hopefully can be easily converted to arguments to Windows’
CreateFile()).

It also defines bitwise manipulations on open_modes, so they can be
used as flags.

Lastly, it defines the default permissions for newly created files
(equivalent to 0600).