Require boost::filesystem on Windows
The Filesystem TS implementation included with GCC6 doesn't seem to ship
on Windows, but having the UTF-16 handling of paths would be really
nice. So (for now) we'll use Boost.Filesystem. (We might be able to
avoid this by using the existing Mac workaround and aliasing fs::path to
std::u16string; I'll have to try it.)