Use std::experimental::filesystem implementation on MSVC
MSVC is annoying, because the header is called <filesystem> but the
namespace is still std::experimental, which breaks the detection logic
we had before. So hard code it for now, this will have to be updated
when MSVC uses namespace std::filesystem (or maybe we can jsut get rid
of the io_std stuff then).