SeekableStreams shouldn't need an offset_type typedef
Use decltype to work out the offset type of a SeekableStream based on
the return value of seek(), rather than relying on a typedef being
present.

This means that basic_adaptor can conditionally be Seekable based on
the member functions its parent has, without requiring either template
gymnastics to conditionally present a member type or a traits class.