Add benchmark from Chistian Adam
A file copy benchmark taken from

https://cristianadam.eu/20160410/c-plus-plus-i-slash-o-benchmark/

and modified to include Modern.

Results: as expected, Modern is a tiny bit slower than Posix, and a
tiny bit faster than stdio and iostreams with libstdc++. With libc++,
Posix is fastest, Modern and stdio are a fraction slower, and iostreams
is very slow indeed.

I’m not sure about the licence for this code as it’s not explicitly
stated on the blog, but I’ll find out.