Add io::stream_position class
For added type safety, separate the notion of “offset” (a distance from
the current file cursor) from that of “position” (which is always the
offset from the beginning of the file). This is roughly the same as the
distinction between a time_point and a duration in the chrono library.

A SeekableStream’s seek() method must return a type which is a model of
StreamPosition. StreamPosition’s requirements are:

  * Is Regular and totally ordered, per the Ranges TS definitions
  * Has a nested typedef offset_type, which is a signed integral type
  * Is (possibly explicitly) constructible from offset_type
  * Has a const member function offset_from_start(), returning
offset_type
  * Has member operator+=/-=(offset_type), returning T&
  * Has non-member operator+/-(T, T::offset_type), returning a new T
  * Has non-member operator-(T, T) returning T::offset_type

The class ion::stream_position<OffT> is an implementation meeting the
above requirements.