Implement Seekable for buffered streams
This is a pretty dumb implementation: if a seek occurs in a reader, we
just dump the cache and pass the seek on to the next layer; in a
writer, we flush the cache then move on.

A smarter implementation would check to see whether the seek is within
the cached area, and if so, perform no IO. But that gets tricky,
especially when the read and write caches might be different sizes.