buffered_read_stream: Add read_next()/peek_next()
These functions return the next byte in the stream, avoiding the
overhead of constructing a buffer calling reading a single byte into it.

Makes iterating over a buffered stream around a hundred times faster,
which is actually kind of annoying, because io::read() shouldn’t be
that slow…

The error_code and non-error-code overloads return different types,
which I’m not altogether happy about, and might change later.